[xcode14.3] Bump api references for .NET 8 and legacy Xamarin. (#19436)
The released commit for .NET 8 isn't the exact same commit as the one
referenced here, because the build.zip file wasn't produced in that commit -
this was fixed, so it's a few commits later. There were no API changes between
these commits though, so the API references are identical.

Backport of #19434.