[Tests] Fix CALayer tests on Sonoma.
Sonoma is picker now with the way we create a CoreGraphics.CGImage,
meaning that if we do not pass the correct flags, the OS will return a
nil object (might be a good idea to move away from constructors to
factory methods).

Before this change we would get the following errors:

* CAKeyFrameAnimation_ValuesTests: System.Exception : Could not initialize an instance of the type 'CoreGraphics.CGImage': handle is null.
It is possible to ignore this condition by setting ObjCRuntime.Class.ThrowOnInitFailure to false.

* CALayer_ValuesTests: System.Exception : Could not initialize an instance of the type 'CoreGraphics.CGImage': handle is null.
It is possible to ignore this condition by setting ObjCRuntime.Class.ThrowOnInitFailure to false.