[Generator] Allow to have a way to attach to the generator.
As soon as we moved to have a build of the generator that used dotnet
we opened the door to a number of developer life improvements. The very
first one of those is the poissibility to attach the dotnet debugger to
the bgen process and that way be able to debug.

In order to do that, this PR has changed two small things:

1. Added code in the main method of the generator that will block the
   tool until a debugger is attached.
2. Changed the csproj to add the previously mentioned code only when
   the enviroment variable XAMMACIOS_DEBUGGER is set. This way the
   code does not reach our customers.

A README has been added explaining how to debug the processes via Visual
Studio.