[CI] Ensure that the new dependencies are correcly managed.
Once we move to the configure stage we have the situation in which we
cannot longer have a simple condition to decide if we are preparing a
released based in a single previos stage.

The issue we have is that templates cannot be used to create condition
expressions and the expressions API from AZP does nor provide a map
function (would have been ideal to do some mapping from the depends on
object to an array).

The way we are going to fix the template is by taking a collection of
stages we depend on, as we did, plus the name of the stage that builds
the pacakges, that way we ensure that the condition will continue to
work.