[release/8.0.1xx] [bgen] Fix assembly comparison. Fixes #19612. (#19639)
This fixes a regression in .NET 8, where we changed the temporary assembly name when
building using a project file / MSBuild - we started compiling the temporary binding
code in MSBuild instead of in the generator, and in the process we changed the name
of the temporary assembly. This broke logic in bgen that compared the assembly name
to check if a given type is from the temporary assembly or not.

Fix this by checking the actual temporary assembly instead of the name of the assembly
instead.

Fixes https://github.com/xamarin/xamarin-macios/issues/19612.

Backport of #19619.