[bgen] Adjust logic to handle protocol-backed events.
Previously we used the presence of the [Protocolize] attribute to detect
Delegate properties returning protocols that should be specially handled if
the delegate in question supported events (in particular the Delegate property
would get additional code to ensure the developer didn't use both the event
pattern and the delegate pattern, because then one or the other wouldn't work
properly).

However, the presence of the [Protocolize] attribute does not take into
account that there's another way to declare that a Delegate property uses a
protocol: the property could use the protocol directly as the property type,
without a [Protocolize] attribute.

So change the logic to detect either the [Protocolize] attribute or an actual
protocol.

This is a slight breaking change in that we'll now add the special code to
more Delegate properties, and we'll throw an exception if the developer uses
both the event pattern and the delegate pattern in more cases. In theory this
should be a good thing (we're detecting more broken developer code), and if it
turns out to be an incorrect check, developers can opt out of this extra check
anyway.