[GameController] Add missing API for XCode15. Fixes #15725
This PR adds the missing bindings and fixes #15725. What the issue reported is incorrect,
the problem with the API is that it uses Swift enumerators that have a String RawValue.
The way to fix this in our bindings is to use the string values that are **PRESENT** in the
headers.