[dotnet] Call the _CreateAssetPackManifest target during the build. Fixes #19669. (#19681)
We're calling the _CreateAssetPackManifest target during the build for legacy
Xamarin apps, but somehow this seems to have been skipped over when
implementing .NET support.

The reason this has not showed up before is that it requires:

* OnDemand resources.
* An AdHoc provisioning profile (and a distribution certificate).

The last part makes it rather complicated to write a unit test, so this has
been verified manually.

Fixes https://github.com/xamarin/xamarin-macios/issues/19669.