[release/8.0.1xx] Fix sending MidiPacket when MidiPacket has been created using a point. (#19716)
Fix sending MidiPacket when MidiPacket has been created using a pointer to a byte[] rather than a byte[] passed as a reference.
MidiPacket.bytes is intiailized to a Array.Empty<byte> so it is never null, and so packet.BytePointer is never used.
We therefore need to check for ByteArray.Length being 0.
Midipacket.bytes is never 0 when intiailized to a 0 length byte[] as that constructor checked for 0 length byte[]

Backport of #18981


---------

Co-authored-by: James Low <j@jameslow.com>