[Tests] Make the test FindMissingObsoleteAttributes usable.
The ToString of a named tuple does not make sense for a human, this
tests, while it does find issues with the API, it does not help the
developer to find the real reason.

This PR has 2 important fixes.

1. Add a generic constrain to the Assert.Failures. The developer that
   wrote the code got lucky and always used T that are IComparable. I
   mean lucky, because the OrderBy from Linq needs T to be IComparable,
   and I am surprise we never got a rutime error with the test.
2. Use a record type with a reasonable ToString.

The following gist contains a diff between the output with the fix and
without it.

https://gist.github.com/mandel-macaque/ee1adab1bf0568a6e95d47e7f3b7970a

The diff is:

Old code:

```
AVFoundation.AVAudioSessionRouteSharingPolicy AVFoundation.AVAudioSessionRouteSharingPolicy::LongForm: AVFoundation.AVAudioSessionRouteSharingPolicy AVFoundation.AVAudioSessionRouteSharingPolicy::LongForm is obsoleted on MacOSX but not on iOS, TVOS, MacCatalyst
```

New code:
```
(AVFoundation.AVAudioSessionRouteSharingPolicy AVFoundation.AVAudioSessionRouteSharingPolicy::LongForm, AVFoundation.AVAudioSessionRouteSharingPolicy AVFoundation.AVAudioSessionRouteSharingPolicy::LongForm, Cecil.Tests.OSPlatformAttributes[], Cecil.Tests.OSPlatformAttributes[])
```