[mmp] Add PreserveSmartEnumConversionsSubStep to the linker sub steps. Fixes #19712.
Looks like this was substep was only added to mtouch, never to mmp, so let's fix that.

This also makes it possible to remove a test that was only there for the
linker to keep some required methods so that the test app would build
correctly.

Fixes https://github.com/xamarin/xamarin-macios/issues/19712.