[UIKit] Fix breaking change in UITraitChangeObservable. (#19886)
The breaking change happened in fc0d09424e85fc30d2a88816b7c20b265e1f210f -
which was a fix to privatize APIss that weren't supposed to be public.
Unfortunately it's too late, because we've already released these APIs, so we
have to keep them public to preserve binary compatibility.