[msbuild] Adjust mangling logical name for the CreateEmbeddedResource, PackLibrary and UnpackLibrary tasks.
In order to make packing and unpacking work on Windows, we now:

* Encode both \ and / the same way, they're both path separator characters.
* Decode either into Path.DirectorySeparatorChar.

This is a potential breaking change: it won't be possible to have a resource
with a LogicalName that contains a backslash. I've never seen an app where
files would have a name containing a backslash, so this is unlikely to be a
problem.