[msbuild] Don't try to bundle the BouncyCastle.Crypto library anymore. (#20044)
A recent bump in the HotRestartVersion seems to be to a version of Hot Restart
that doesn't need/provide the BouncyCastle.Crypto library.

Unfortunately due to another issue, the bump didn't actually bring in the
updated Hot Restart version (the required NuGet packages weren't published to
a feed we subscribe to, so we silently [1] got a different version of the Hot
Restart packages). This means our build didn't break in the actual bump, but
only when the corresponding Hot Restart package versions were published to the
designated feeds (i.e. yesterday).

[1]: We got a warning, but no build error. #20043 fixes this problem.