[UIKit] Remove presnippet for UITextView.TypingAttributes. Fixes #12709.
Apparently, in earlier versions of iOS, returning an empty dictionary from
UITextView.TypingAttributes would crash, so we added a workaround where we'd
return null instead of an empty dictionary.

However, I can't reproduce any such crashes anymore, and this workaround is
causing other problems, so remove it.

But do this in a backwards compatible way (just in case the removal ends up
crashing apps after all):

* Add a new TypingAttributes2 binding with the behavior we want (a better name
  wouldn't be unwelcome!)
* Mark the existing TypingAttributes as obsolete, pointing to
  TypingAttributes2.
* Make the TypingAttributes property behave as we want in XAMCORE_5_0, and at
  that point  mark TypingAttributes2 as obsolete and point back to
  TypingAttributes.
* Remove TypingAttributes2 in XAMCORE_6_0.

Note: the presnippet attribute to avoid the crash was introduced in 2013:
https://github.com/xamarin/maccore/commit/5adb7c8608934413768f000a04f7dd6fddef7eca,
and our current minimum iOS version (iOS 11), was launched in 2017, so most
likely (aka hopefully) the crash was fixed in an iOS version earlier than the
earliest we support.

Fixes https://github.com/xamarin/xamarin-macios/issues/12709.