[msbuild] Use MetadataLoadContext in the UnpackLibraryResources task. (#20017)
This avoids the Mono.Cecil dependency, and also MetadataLoadContext is both
faster and better maintained.

A temporary downside is that all the reference assemblies will have to be
copied to the remote Mac for remote builds, but this will be fixed in a future
change where we'll execute this task mostly locally on Windows, and we'll
only copy the unpacked resources to the Mac.