[devops] Don't allow re-running just the Windows tests job.
Don't allow re-running just the Windows tests job, because the Mac that was
reserved for us isn't reserved anymore.

Instead all the jobs must be re-run.

Accomplish this by storing the current BuildId in a file on the Mac bot, and
then verifying that it's the expeted BuildId from the test step from Windows.