[tests] Verify that we have a remote connection in tests that require a remote connection. (#20114)
By default, the build will just continue if a remote connection couldn't be
established. For tests that require a remote connection, this often results in
weird errors later on in the build, so improve this by immediately verifying
that we were able to establish a remote connection and telling the developer
where to look for more information.

---------

Co-authored-by: Alex Soto <alex@alexsoto.me>