[GitOps] Add an opt-out for automated comments for need-* labels. (#20170)
Currently a descriptive comment is automatically added to an issue when either
the 'need-info' or the 'need-repro' label is added to an issue, requesting
more info or a test project, respectively, and then when the reporter (or
anyone else) comments something, the `need-info`/`need-repro` label is removed
and a `need-attention` label is added.

However, sometimes this can get repetitive if the reporter replies frequently,
or someone else steps in and answers questions or asks unrelated questions,
and we have to add back the `need-info` or `need-repro` every time.

So add an opt-out: if the label `no-auto-reply` is set, then don't add that
descriptive comment.

The label is single-use: it'll be removed when a `need-info` or `need-repro`
labels is added.