[devops] Don't set any pending/success/failure statuses. (#20184)
DevOps automatically sets statuses as the pipeline progresses, so we
don't have to.

This fixes a problem where we'd have multiple statuses for each thing we
do:

![Screenshot 2024-02-22 at 18 09
42](https://github.com/xamarin/xamarin-macios/assets/249268/c49c3a08-9f7e-45d9-a662-64022b120643)

Note how the macOS .NET tests have two statuses, one pending and one
failure. This is because it failed once (thus the failure), and was then
rerun (thus pending). This has a few issues:

1. The statuses don't match... which can be confusing. Are we running or
did we fail? YES! 🤔
2. We have more than one status for each thing, which is redundant.

So just don't add our own statuses, because Azure DevOps will.

Note that we still add statuses with package links.